Move system specific dlopen related stuff to utils.
