+++ thorin/CMakeLists.txt
@@ -86,6 +86,8 @@ add_library(libthorin
     util/bitset.h
     util/cast.h
     util/container.h
+    util/dlopen.cpp
+    util/dlopen.h
     util/hash.cpp
     util/hash.h
     util/indexmap.h
